Rename Point to Vertex
