+++ locales/geometry/Vertex/en.json
