+++ locales/geometry/Vertex/ja.json
