+++ locales/geometry/Vertex/list.js
